# backend/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "chi (net/http) + pgx (PostgreSQL) + Redis дээр суурилсан Clean Architecture бүхий Go backend. Нээлттэй эхийн snykk/go-rest-boilerplate (MIT, зохиогч Najib Fikri)-ээс үүсэлтэй; HTTP давхаргыг chi, өгөгдлийн давхаргыг pgx руу хөрвүүлсэн.",
        "title": "eID based AI enabled Government Template Platform V3.0 API",
        "termsOfService": "https://github.com/snykk/go-rest-boilerplate",
        "contact": {
            "name": "eID based AI enabled Government Template Platform V3.0",
            "url": "https://github.com/snykk/go-rest-boilerplate"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/auth/eid/poll": {
            "post": {
                "description": "session_id-ийн төлвийг IdP-ээс long-poll-оор (≤25с) асууна. state нь RUNNING/COMPLETE/EXPIRED/REFUSED. COMPLETE үед identity-аар хэрэглэгчийг бүртгэж/шинэчилж, access+refresh токен хосыг /login-той ижил хэлбэрээр буцаана.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "eID session-ийн төлвийг асуух (long-poll)",
                "parameters": [
                    {
                        "description": "eID session id",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/template_internal_http_datatransfers_requests.EIDPollRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Session state (with tokens if COMPLETE)",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/template_internal_http_handlers_v1.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/template_internal_http_datatransfers_responses.EIDPollResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Malformed JSON body or missing session_id",
                        "schema": {
                            "$ref": "#/definitions/template_internal_http_handlers_v1.BaseResponse"
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/template_internal_http_handlers_v1.BaseResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to reach eID provider",
                        "schema": {
                            "$ref": "#/definitions/template_internal_http_handlers_v1.BaseResponse"
                        }
                    }
                }
            }
        },
        "/auth/eid/start": {
            "post": {
                "description": "Гадаад eID identity provider дээр QR/deep-link нэвтрэлтийг эхлүүлж, клиент харуулах session мэдээллийг (session_id, device_link_url, verification_code, expires_at) буцаана. Дараа нь /auth/eid/poll руу session_id-г дамжуулж төлвийг асууна.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "eID нэвтрэлт эхлүүлэх (QR / deep-link)",
                "responses": {
                    "200": {
                        "description": "eID session started",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/template_internal_http_handlers_v1.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/template_internal_http_datatransfers_responses.EIDStartResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Failed to reach eID provider",
                        "schema": {
                            "$ref": "#/definitions/template_internal_http_handlers_v1.BaseResponse"
                        }
                    }
                }
            }
        },
        "/auth/eid/start-id": {
            "post": {
                "description": "Иргэний РД (national_id)-аар нэвтрэлтийг эхлүүлж, тухайн РД-тэй холбоотой бүртгэлтэй төхөөрөмж рүү баталгаажуулах prompt push хийлгэнэ. QR/device_link шаардлагагүй тул зөвхөн session_id, verification_code, expires_at буцна. Дараа нь /auth/eid/poll руу session_id-г дамжуулж төлвийг асууна.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "eID нэвтрэлт эхлүүлэх (РД-аар push)",
                "parameters": [
                    {
                        "description": "Иргэний РД",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/template_internal_http_datatransfers_requests.EIDStartByNationalIDRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "eID session started",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/template_internal_http_handlers_v1.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/template_internal_http_datatransfers_responses.EIDStartResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Malformed JSON body or missing national_id",
                        "schema": {
                            "$ref": "#/definitions/template_internal_http_handlers_v1.BaseResponse"
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/template_internal_http_handlers_v1.BaseResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to reach eID provider",
                        "schema": {
                            "$ref": "#/definitions/template_internal_http_handlers_v1.BaseResponse"
                        }
                    }
                }
            }
        },
...
        "/auth/register": {
            "post": {
                "description": "Хэрэглэгчийн бүртгэлийг идэвхгүй төлөвт үүсгэнэ. Дуудагч нь /auth/login амжилттай болохоос өмнө бүртгэлийг идэвхжүүлэхийн тулд /auth/send-otp + /auth/verify-otp-г дагах ёстой.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Шинэ хэрэглэгч бүртгэх",
                "parameters": [
                    {
                        "description": "Registration payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/template_internal_http_datatransfers_requests.RegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "User created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/template_internal_http_handlers_v1.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/template_internal_http_datatransfers_responses.UserResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Malformed JSON body",
                        "schema": {
                            "$ref": "#/definitions/template_internal_http_handlers_v1.BaseResponse"
                        }
                    },
                    "409": {
                        "description": "Email or username already in use",
                        "schema": {
                            "$ref": "#/definitions/template_internal_http_handlers_v1.BaseResponse"
                        }
                    },
                    "422": {
                        "description": "Validation error (per-field detail in data.errors)",
                        "schema": {
                            "$ref": "#/definitions/template_internal_http_handlers_v1.BaseResponse"
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "$ref": "#/definitions/template_internal_http_handlers_v1.BaseResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "template_internal_http_datatransfers_requests.EIDPollRequest": {
            "type": "object",
            "required": [
                "session_id"
            ],
            "properties": {
                "google_link_token": {
                    "description": "GoogleLinkToken нь Google-ээр эхний удаа нэвтэрсэн хэрэглэгч eID-ээр\nбаталгаажуулж байгаа үед л ирнэ (сонголттой).",
                    "type": "string"
                },
                "session_id": {
                    "type": "string"
                }
            }
        },
        "template_internal_http_datatransfers_requests.EIDStartByNationalIDRequest": {
            "type": "object",
            "required": [
                "national_id"
            ],
            "properties": {
                "callbackUrl": {
                    "description": "CallbackUrl (сонголт): SAME-DEVICE (утасны browser) үед \u003corigin\u003e/auth/eid/callback; хоосон\nбол CROSS-DEVICE (desktop). Backend force-normalize хийнэ.",
                    "type": "string"
                },
                "national_id": {
                    "type": "string"
                }
            }
        },
...
        "template_internal_http_datatransfers_requests.RegisterRequest": {
            "type": "object",
            "required": [
                "email",
                "first_name",
                "last_name",
                "password",
                "username"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "maxLength": 50
                },
                "first_name": {
                    "description": "нэр (монгол)",
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 1
                },
                "first_name_en": {
                    "description": "нэр (англи)",
                    "type": "string",
                    "maxLength": 50
                },
                "last_name": {
                    "description": "овог (монгол)",
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 1
                },
                "last_name_en": {
                    "description": "овог (англи)",
                    "type": "string",
                    "maxLength": 50
                },
                "password": {
                    "type": "string",
                    "maxLength": 72,
                    "minLength": 12
                },
                "username": {
                    "type": "string",
                    "maxLength": 25,
                    "minLength": 3
                }
            }
        },
...
        "template_internal_http_datatransfers_responses.EIDCertificate": {
            "type": "object",
            "properties": {
                "issuer": {
                    "type": "string"
                },
                "key_type": {
                    "type": "string"
                },
                "not_after": {
                    "type": "string"
                },
                "not_before": {
                    "type": "string"
                },
                "serial": {
                    "type": "string"
                }
            }
        },
...
        "template_internal_http_datatransfers_responses.EIDInfo": {
            "type": "object",
            "properties": {
                "certificate": {
                    "$ref": "#/definitions/template_internal_http_datatransfers_responses.EIDCertificate"
                },
                "civil_id": {
                    "type": "string"
                },
                "document_number": {
                    "type": "string"
                },
                "kyc_level": {
                    "description": "сертификатын түвшин",
                    "type": "string"
                },
                "national_id": {
                    "description": "регистрийн дугаар",
                    "type": "string"
                }
            }
        },
...
        "template_internal_http_datatransfers_responses.EIDPollResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "eid": {
                    "description": "EID нь eID-ээр нэвтэрсэн хэрэглэгчийн identity + сертификатын мэдээлэл.\nНууц үгээр бүртгүүлсэн хэрэглэгчид nil (omitempty).",
                    "allOf": [
                        {
                            "$ref": "#/definitions/template_internal_http_datatransfers_responses.EIDInfo"
                        }
                    ]
                },
                "email": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "first_name_en": {
                    "type": "string"
                },
                "full_name": {
                    "type": "string"
                },
                "full_name_en": {
                    "type": "string"
                },
                "google": {
                    "description": "Google нь холбогдсон Google account-аас хадгалсан профайл. Google\nхолбоогүй хэрэглэгчид nil (omitempty).",
                    "allOf": [
                        {
                            "$ref": "#/definitions/template_internal_http_datatransfers_responses.GoogleInfo"
                        }
                    ]
                },
                "id": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                },
                "last_name_en": {
                    "type": "string"
                },
                "refresh_token": {
                    "type": "string"
                },
                "role_id": {
                    "type": "integer"
                },
                "state": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "template_internal_http_datatransfers_responses.EIDStartResponse": {
            "type": "object",
            "properties": {
                "device_link_url": {
                    "type": "string"
                },
                "expires_at": {
                    "type": "string"
                },
                "session_id": {
                    "type": "string"
                },
                "verification_code": {
                    "type": "string"
                }
            }
        },
...
        "template_internal_http_datatransfers_responses.GoogleInfo": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "email_verified": {
                    "type": "boolean"
                },
                "linked_at": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "picture": {
                    "type": "string"
                }
            }
        },
...
        "template_internal_http_datatransfers_responses.UserResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "eid": {
                    "description": "EID нь eID-ээр нэвтэрсэн хэрэглэгчийн identity + сертификатын мэдээлэл.\nНууц үгээр бүртгүүлсэн хэрэглэгчид nil (omitempty).",
                    "allOf": [
                        {
                            "$ref": "#/definitions/template_internal_http_datatransfers_responses.EIDInfo"
                        }
                    ]
                },
                "email": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "first_name_en": {
                    "type": "string"
                },
                "full_name": {
                    "type": "string"
                },
                "full_name_en": {
                    "type": "string"
                },
                "google": {
                    "description": "Google нь холбогдсон Google account-аас хадгалсан профайл. Google\nхолбоогүй хэрэглэгчид nil (omitempty).",
                    "allOf": [
                        {
                            "$ref": "#/definitions/template_internal_http_datatransfers_responses.GoogleInfo"
                        }
                    ]
                },
                "id": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                },
                "last_name_en": {
                    "type": "string"
                },
                "refresh_token": {
                    "type": "string"
                },
                "role_id": {
                    "type": "integer"
                },
                "token": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "template_internal_http_handlers_v1.BaseResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "request_id": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "/auth/login эсвэл /auth/refresh-аас олгогдсон Bearer хандах токен (access token).",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}